+++ packages/desktop/src/modules/music/shared.ts
@@ -11,7 +11,7 @@ export const getTempDir = async () => {
 }
 
 export const getCachedLyricInfo = async (musicInfo: AnyListen.Music.MusicInfo): Promise<AnyListen.Music.LyricInfo | null> => {
-  let lrcInfo = await workers.dbService.getRawLyric(musicInfo.id)
+  let lrcInfo = await workers.dbService.getPlayerLyric(musicInfo.id)
   // lrcInfo = {} as unknown as AnyListen.Player.LyricInfo
   if (existTimeExp.test(lrcInfo.lyric)) {
     return lrcInfo
@@ -23,24 +23,9 @@ export const saveLyricInfo = async (musicInfo: AnyListen.Music.MusicInfo, info: 
   await workers.dbService.rawLyricSave(musicInfo.id, info)
 }
 
-// TODO rawlrcInfo s2t
 export const buildLyricInfo = async (lyricInfo: AnyListen.Music.LyricInfo): Promise<AnyListen.Music.LyricInfo> => {
   if (appState.appSetting['player.isS2t']) {
-    const tasks = [
-      lyricInfo.lyric ? workers.utilService.langS2T(lyricInfo.lyric) : Promise.resolve(''),
-      lyricInfo.tlyric ? workers.utilService.langS2T(lyricInfo.tlyric) : Promise.resolve(''),
-      lyricInfo.rlyric ? workers.utilService.langS2T(lyricInfo.rlyric) : Promise.resolve(''),
-      lyricInfo.awlyric ? workers.utilService.langS2T(lyricInfo.awlyric) : Promise.resolve(''),
-    ]
-    return Promise.all(tasks).then(([lyric, tlyric, rlyric, awlyric]) => {
-      return {
-        ...lyricInfo,
-        lyric,
-        tlyric,
-        rlyric,
-        awlyric,
-      }
-    })
+    return workers.utilService.lyricS2T(lyricInfo)
   }
   return lyricInfo
 }