+++ packages/view-main/src/modules/lyric/lyric.ts
@@ -4,14 +4,13 @@ import Lyric from '@any-listen/web/lyric-font-player'
 import { clearTimeoutBg, setTimeoutBg } from '@/shared/tools'
 
 import { settingState } from '../setting/store/state'
-import { setLines, setTempOffset, setText } from './store/action'
+import { setLines, setText } from './store/action'
 import type { Line } from './store/state'
 
 let lrc: Lyric | null
 
 export const setOffset = (offset: number) => {
   lrc?.setOffset(offset)
-  setTempOffset(offset)
 }
 
 export const setPlaybackRate = (rate: number) => {
@@ -82,8 +81,6 @@ export const initLyric = () => {
       // console.log(lines) // lines is array of all lyric text
       setLines([...lines])
       setText('', -1)
-      setOffset(offset) // 歌词延迟
-      setTempOffset(0) // 重置临时延迟
     },
     onUpdateLyric(lines: Line[]) {
       setLines([...lines])