+++ packages/view-main/src/modules/lyric/store/commit.ts
@@ -12,9 +12,3 @@ export const setText = (text: string, line: number) => {
   lyricState.line = line
   lyricEvent.lineChanged(text, line)
 }
-export const setOffset = (offset: number) => {
-  lyricState.offset = offset
-}
-export const setTempOffset = (offset: number) => {
-  lyricState.tempOffset = offset
-}